# Optimize `I18n.t`