--- a/lib/i18n.rb
+++ b/lib/i18n.rb
@@ -331,12 +331,11 @@ module I18n
     # keys are Symbols.
     def normalize_keys(locale, key, scope, separator = nil)
       separator ||= I18n.default_separator
+      locale = locale.to_sym if locale
 
-      [
-        *normalize_key(locale, separator),
-        *normalize_key(scope, separator),
-        *normalize_key(key, separator)
-      ]
+      result = [locale]
+      result.concat(normalize_key(scope, separator)) if scope
+      result.concat(normalize_key(key, separator))
     end
 
     # Returns true when the passed locale, which can be either a String or a