--- a/lib/i18n/interpolate/ruby.rb
+++ b/lib/i18n/interpolate/ruby.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 # heavily based on Masao Mutoh's gettext String interpolation extension
 # http://github.com/mutoh/gettext/blob/f6566738b981fe0952548c421042ad1e0cdfb31e/lib/gettext/core_ext/string.rb
 
@@ -10,6 +12,11 @@ module I18n
   INTERPOLATION_PATTERN = Regexp.union(DEFAULT_INTERPOLATION_PATTERNS)
   deprecate_constant :INTERPOLATION_PATTERN
 
+  INTERPOLATION_PATTERNS_CACHE = Hash.new do |hash, patterns|
+    hash[patterns] = Regexp.union(patterns)
+  end
+  private_constant :INTERPOLATION_PATTERNS_CACHE
+
   class << self
     # Return String or raises MissingInterpolationArgument exception.
     # Missing argument's logic is handled by I18n.config.missing_interpolation_argument_handler.
@@ -20,7 +27,8 @@ module I18n
     end
 
     def interpolate_hash(string, values)
-      string.gsub(Regexp.union(config.interpolation_patterns)) do |match|
+      pattern = INTERPOLATION_PATTERNS_CACHE[config.interpolation_patterns]
+      string.gsub(pattern) do |match|
         if match == '%%'
           '%'
         else